+++ dev/jsxgraph/sci_extensions.cljs
@@ -1,0 +1,6 @@
+(ns jsxgraph.sci-extensions
+  (:require [jsxgraph.sci]))
+
+;; ## SCI Customization
+
+(jsxgraph.sci/install!)