+++ src/deps.cljs
@@ -1,2 +1,2 @@
 {:npm-deps
- {"jsxgraph" "1.5.0-rc1"}}
+ {"jsxgraph" "1.5.0"}}